requirements.txt
requests
python-dotenv
termcolor
pandas
openpyxl
PyYAML